User-authentication

+++ controller/contactController.js
@@ -5,26 +5,27 @@ import Contact from '../models/contactModel.js';
 /* 
     @Description : To get all contacts
     @Route : GET /api/contacts
-    @Access : public
+    @Access : PRIVATE
 */
 export const getContact = asyncHandler(async (req, res) => {
-  const contact = await Contact.find()
+  const contact = await Contact.find({user_id:req.user.id})
   res.status(200).send({isSuccess:true, contact});
 });
 
 /* 
     @Description : To createContact
     @Route : POST /api/contacts
-    @Access : public
+    @Access : PRIVATE
 */
 export const createContact = asyncHandler(async (req, res) => {
   const { name, email, phone } = req.body;
+  const userID = req.user.id;
   if (!name || !email || !phone) {
     res.status(400);
     throw new Error("All fields are required");
   }
   const contact = await Contact.create({
-    name,email,phone
+    user_id:userID,name,email,phone
   })
 
   res.status(201).send({isSuccess:true, contact});
@@ -33,11 +34,12 @@ export const createContact = asyncHandler(async (req, res) => {
 /* 
     @Description : To getIndividualContact
     @Route : GET /api/contacts/:id
-    @Access : public
+    @Access : PRIVATE
 */
 export const getIndividualContact = asyncHandler(async (req, res, next) => {
   const userID = req.params.id;
   const contact = await Contact.findById(userID);
+  console.log("🚀 ~ getIndividualContact ~ contact:", contact)
   if (!contact) {
     res.status(404);
     throw new Error(`Contact not found with id ${userID}`);
@@ -47,35 +49,49 @@ export const getIndividualContact = asyncHandler(async (req, res, next) => {
 /* 
     @Description : To updateContact
     @Route : PUT /api/contacts/:id
-    @Access : public
+    @Access : PRIVATE
 */
 export const updateContact = asyncHandler(async (req, res) => {
-  const userID = req.params.id;
-  const contact = await Contact.findById(userID);
+  const ID = req.params.id;
+  const contact = await Contact.findById(ID);
+  const user_id = req.user.id; 
   if (!contact)
   {
     res.status(404);
-    throw new Error("No contact found with id " + userID)
+    throw new Error("No contact found with id " + ID)
   }
-  const updatedContact = await Contact.findByIdAndUpdate(userID, req.body, { new: true });
-  res.status(200).send({isSuccess: true,message: `Contact updated for ID : ${userID}`, updatedContact });
+  console.log("CHECK",contact._id.toString(), user_id)
+  if (contact.user_id.toString() !== user_id)
+  {
+    res.status(403);
+    throw new Error("not AUTHORIZED to perform this action ")
+  }
+    
+  const updatedContact = await Contact.findByIdAndUpdate(ID, req.body, { new: true });
+  res.status(200).send({isSuccess: true,message: `Contact updated for ID : ${ID}`, updatedContact });
 });
 
 /* 
     @Description : To deleteContact
     @Route : DELETE /api/contacts/:id
-    @Access : public
+    @Access : PRIVATE
 */
 export const deleteContact = asyncHandler(async (req, res) => {
   const userID = req.params.id;
   const contact = await Contact.findById(userID);
-  console.log("🚀 ~ deleteContact ~ contact:", contact._id)
   if (!contact)
   {
     res.status(404);
     throw new Error(`Contact not found`)
   }
+
+  if (contact.user_id.toString() !== req.user.id)
+    {
+      res.status(403);
+      throw new Error("Not AUTHORIZED to perform this action ")
+  }
+  
   const deletedContact = await Contact.findByIdAndDelete(userID);
 
-  res.status(200).send({isSuccess: true, message: `Contact Deleted with ID: ${userID}`, contact});
+  res.status(200).send({isSuccess: true, message: `Contact Deleted with ID: ${userID}`, deletedContact});
 });
